+++ jodd-mail/src/main/java/jodd/mail/Pop3Server.java
@@ -46,23 +46,11 @@ public class Pop3Server extends MailServer<ReceiveMailSession> {
 	 */
 	protected static final int DEFAULT_POP3_PORT = 110;
 
-	public Pop3Server(
-			final String host,
-			final int port,
-			final Authenticator authenticator,
-			final File attachmentStorage,
-			final int timeout,
-			final boolean strictAddress,
-			final boolean debugMode
-			) {
-		super(
-			host,
-			port == -1 ? DEFAULT_POP3_PORT : port,
-			authenticator,
-			attachmentStorage,
-			timeout,
-			strictAddress,
-			debugMode);
+	public Pop3Server(final Builder builder) {
+		super(builder, DEFAULT_POP3_PORT);
+	}
+	protected Pop3Server(final Builder builder, final int defaultPort) {
+		super(builder, defaultPort);
 	}
 
 	@Override